Expand scrollbar input area for song select carousel
